体温カラムの方変更と体温入力バリデーションを追加

+++ database/migrations/2020_11_03_053402_create_conditions_table.php
@@ -16,7 +16,7 @@ class CreateConditionsTable extends Migration
         Schema::create('conditions', function (Blueprint $table) {
             $table->id();
             $table->integer('user_id');
-            $table->integer('body_temperture');
+            $table->double('body_temperture',4,1);
             $table->integer('nail');
             $table->integer('belly');
             $table->integer('rough_hands');
